tests/wpt/web-platform-tests/shadow-dom/focus-navigation/focus-navigation-scroller-tabindex.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="resources/shadow-dom.js"></script>
<script src="resources/focus-utils.js"></script>

<style>
  div.scroll {
    overflow: auto;
    width: 20em;
    height: 5em;
    display: block;
  }

  div.long {
    width: 40em;
  }
</style>

<p>Tests for moving focus on scrollable element by pressing tab key.</p>
<button id="start">START</button>

<p>Where element has positive tabindex.</p>
<div id="scroller-positive" class="scroll" tabindex="0">
  <div class="long"></div>
  <button id="A">BUTTON A</button>
  <div>Text</div>
</div>

<div id="scroller-positive-2" class="scroll" tabindex="0">
  <div class="long"></div>
  <div>Text</div>
</div>

<p>Where element has no tabindex.</p>
<div id="scroller-none" class="scroll">
  <div class="long"></div>
  <button id="B">BUTTON B</button>
  <div>Text</div>
</div>

<div id="scroller-none-2" class="scroll">
  <div class="long"></div>
  <div>Text</div>
</div>

<p>Where element has negative tabindex.</p>
<div id="scroller-negative" class="scroll" tabindex="-1">
  <div class="long"></div>
  <button id="C">BUTTON C</button>
  <div>Text</div>
</div>

<div id="scroller-negative-2" class="scroll" tabindex="-1">
  <div class="long"></div>
  <div>Text</div>
</div>

<button id="end">END</button>

<script>

promise_test(async () => {
  let elements = [
    'start',
    'scroller-positive',
    'A',
    'scroller-positive-2',
    'B',
    'scroller-none-2',
    'C',
    'end',
  ];

  await assert_focus_navigation_forward(elements);
  elements.reverse();
  await assert_focus_navigation_backward(elements);
}, 'Should focus on scroller depending on tabindex value.');

</script>
